+++ appeals/pdf-service-api/src/mappers/appellant-case/appellant-case.mapper.js
@@ -2,6 +2,7 @@ import { additionalDocumentsSection } from './sections/additional-documents.sect
 import { appealDetailsSection } from './sections/appeal-details.section.js';
 import { appellantDetailsSection } from './sections/appellant-details.section.js';
 import { applicationDetailsSection } from './sections/application-details.section.js';
+import { beforeYouStartSection } from './sections/before-you-start.section.js';
 import { siteDetailsSection } from './sections/site-details.section.js';
 import { uploadDocumentsSection } from './sections/upload-documents.section.js';
 
@@ -10,6 +11,7 @@ export default function mapAppellantCaseData(templateData) {
 	return {
 		details: { appealSite, appealReference, localPlanningDepartment },
 		sections: [
+			beforeYouStartSection(templateData),
 			appellantDetailsSection(templateData),
 			siteDetailsSection(templateData),
 			applicationDetailsSection(templateData),
